A4: Add logging patch and sequence assignments
The file `enable_logging.patch` is a patch to apply to the code from class.
It makes the driver log the order of every operation performed on it.

The folder `sequences` contains a file for each student with their sequence in it.
Once you apply the logging patch, your job is to write a userspace program that
performs the nessecarily operations on the special device file `/dev/basic_char`
to make the driver log printed to the ring buffer (dmesg) match your assigned sequencer.

Signed-off-by: Charles Mirabile <cmirabil@redhat.com>